Chaange gross. pnl to float to account for other traded instruments

+++ trading_portfolio/pnl_calendar/models.py
@@ -6,6 +6,6 @@ from django.db import models
 class Trades(models.Model):
     trade_id = models.CharField(primary_key=True)
     trade_date = models.DateTimeField()
-    gross_pnl = models.IntegerField()
+    gross_pnl = models.FloatField()
     fee = models.FloatField()
     net_pnl = models.FloatField()
